+++ lib/extension/gallery_assets.dart
@@ -1,0 +1,13 @@
+/*
+ * @Author: lipeng 1162423147@qq.com
+ * @Date: 2023-09-29 11:08:34
+ * @LastEditors: lipeng 1162423147@qq.com
+ * @LastEditTime: 2023-09-29 11:08:53
+ * @FilePath: /phoenix_gallery/lib/extension/gallery_assets.dart
+ * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
+ */
+import 'package:phoenix_base/phoenix.dart';
+
+extension GalleryAssets on BaseAsset {
+  static const String iconUpArrow = 'icons/icon_up_arrow.png';
+}