+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+<!--
+ * @Author: lipeng 1162423147@qq.com
+ * @Date: 2023-09-24 18:00:03
+ * @LastEditors: lipeng 1162423147@qq.com
+ * @LastEditTime: 2023-09-29 11:10:38
+ * @FilePath: /phoenix_gallery/CHANGELOG.md
+ * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
+-->
 ## 1.0.0
 
 * [新建] 企业级基础组件:Gallery
+
+## 1.0.1
+
+* [更新] resource path